mtk-vcodec: Use specific MT6768 headers as much as possible
Some needed functions require specific MT6768 definitions due to somme reason.
Add `if else` conditions to guard some definitions, so
if we build for MT6768, then Linux build system will use
the said definitions as much as possible.

This fixes multiple errors that was related to incompatible functions
usage, or something like that.

Probably needs to be double-checked.

Signed-off-by: Hendra Manudinata <hendra@manudinata.me>